improving providers and hooks
